feat: container page table component
Switch the Containers page to use the Table component. This is a final piece
of technical debt adopting the Table component, and gives a few advantages:
- Table expands/collapses to fit the display width (used to extend off the
  right edge with a scrollbar).
- Sorting of name column to start (wanted to focus on core aspects and get
  this in before sorting other columns).
- Consistent table UI with other pages.
- Basic support for light mode.

The containers table included ContainerGroupInfoUI parents of three types
(pods, compose groups, and standalone containers), all with ContainerInfoUI
children. To avoid UI issues the parent ContainerGroupInfoUI was skipped
when drawing standalone containers (i.e. it only draws the child container
object in this one case).

Since the Table component doesn't support this unique behaviour, the least
invasive/simplest option was to create a derived list where standalone
containers are mapped to actual containers.

Found an issue where Svelte was updating constantly; turns out this was
an existing bug in the timeout when no containers are running.

Ignore changes to Table - reviewed separately under #7587.

Fixes #4843.

Signed-off-by: Tim deBoer <git@tdeboer.ca>