chore: fix name spacing in kubernetes dashboards
### What does this PR do?

The dashboard for ingress with a lot of information is "squished" and
hard to read.

This PR changes the default width of ingress/route table as well as
deployments to "1fr" which matches the current
Services list formatting.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

Visual change, no test really required TBH, check `yarn watch` and view
the tables for changes.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>