chore: add unsubscribe for k8s deployments (#7337)
### What does this PR do?

Adds the missing unsubscribe for deployments.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop/issues/7336

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

Tests cover the feature. Ex. `(window as
any).window.kubernetesUnregisterGetCurrentContextResources = () =>
Promise.resolve(undefined);` is added. If that was not there, it would
fail (it will try and call unsubscribe).

Signed-off-by: Charlie Drage <charlie@charliedrage.com>