chore: tooltip simplification
Changes
- Move 'content' back to be the default slot, since content must always be there and is naturally the visible child.
- Make the outer box consistent - i.e. you are welcome to customize the _content_ of tooltips, but the component creates the frame and bg color.
- In additional to the 'tip' slot, allow 'tip' to be a property, so in the simple cases you can still do:
```
<Tooltip right tip="some tip">
  <something/>
</Tooltip>
```
and when you want to totally customize the tip content:
```
<Tooltip right>
  <something/>
  <svelte:fragment slot="tip">tip content goes here</svelte:fragment>
</Tooltip>
```

If someone accidentally supplies both, the simple tip is used. Having two if-statements is minor duplication, but it is more obvious and avoids extra spaces in the html.

I know this undoes a bunch of change and 'tip can either be property or slot' is a bit different (could use different names?), but I think this handles the simple/normal case as easy as before and has a single 'frame', while still allowing for enough customization.

Fixes #7435.

Signed-off-by: Tim deBoer <git@tdeboer.ca>