chore: label component
We have labels in several table columns: a rounded bg with a dot or icon + text.
(re:naming - we already have 'badges', and these are like github labels)
The existing labels have copied and inconsistent styling, so this creates
a new Label component and reuses it for these cases. The name property
provides the text and icon or dot goes in the slot, with optional tooltip.

Image/pod had duplicate/unnecessary styling, removed rounded and text-sm.
Removed styling checks from downstream tests.

Part of #7348.

Signed-off-by: Tim deBoer <git@tdeboer.ca>