chore: add tooltip to show load balancer IP for service
### What does this PR do?

Adds a tooltip with the IP so you can easily see what the load balancer
IP is.

As a follow up PR, it'd be nice to copy this to clipboard.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop/issues/7426

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

Create a load balancer and hover over the badge.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>