chore: fix table sorting
### What does this PR do?

When sorting, we are sorting the columns on the onMount, however, we are
not waiting for data (tick()) as well as not sorting it afterwards
(sortImpl()).

This PR fixes it by sorting the table on launch.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Fixes https://github.com/containers/podman-desktop/issues/7475

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

View the pages and see they are ordered now.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>