chore: support different parent/child types in tables
A previous PR added support for having child items in tables, but the type
of the parent and child objects is the same.

For the Containers page, the parents are always ContainerGroupInfoUI and
the children are always ContainerInfoUI. This change just adds an optional
second type (identically to how we did rendering types in Column) so that
you can declare the different types naturally as:
  `ContainerGroupInfoUI, ContainerInfoUI`
instead of having to do:
  `ContainerGroupInfoUI | ContainerInfoUI`
and deal with typing issues in the client code.

Part of #4843.

Signed-off-by: Tim deBoer <git@tdeboer.ca>