chore: fix offset icons
I noticed that the icon on Kubernetes node status isn't centered - the svg is
cropped a little on the left so everything is a few pixels to the left.

When fixing this I noticed that (for now) it's using the same icon as Settings,
which has the same problem - and fixing that showed that the Accounts nav item
even had a left border in order to compensate and line up.

As far as I know none of this is intentional and should be fixed so that all
the items in the nav line up, and node status icons are fixed.

Fixes #7556.

Signed-off-by: Tim deBoer <git@tdeboer.ca>