chore: add overflow for k8s node role
### What does this PR do?

Adds the "overflow" for the column so that the text does not get
squished.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A, forgot to add it.

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

N/A, small change. Does not "squish" the name / hide the text. See
above screenshot.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>