chore: switch to using name for expansion id
When I added parent/child groups to Table it felt wrong to use an object's
name property as the unique id for knowing what is expanded/collapsed, and
all the objects I was looking at had id properties so I used that instead.

However, when porting the Containers page we have already been using the
name property, it is already required to be unique for things like
aria-labels, and some objects don't have ids. To keep things simpler,
easier to port, and have only one required property I'm switching to name.

Signed-off-by: Tim deBoer <git@tdeboer.ca>