chore: move status colors to registry
Moves status colors to the color registry.

Most of this is just porting; the one addition I had to make is
StatusIcon was using a hard-coded color for things that aren't
running - which is *not* the same color as stopped/exited, etc.
To avoid changing colors elsewhere, for now I've just defined a
new 'not-running' status color, which we may want to revisit
with UX later.

Fixes #7581.

Signed-off-by: Tim deBoer <git@tdeboer.ca>