chore: light mode for nav section
NavItem had been moved to light mode, but NavSection (the expandable
sections for extensions and Kubernetes) was still picking up default
colors, which happened to pick up white even in dark mode. This just
sets the text color explicitly using the color from the registry, and
adds the missing hover color too.

Compared to regular nav items there is still a 'missing' hover bg, but
since it isn't defined how/if that should work for sections I'm not
adding it here.

Signed-off-by: Tim deBoer <git@tdeboer.ca>