chore: use color var in index.html
We should not hardcode a default text colour to white, especially in light
mode. This is the (temporary) cause of many issues like #7708 where we
hadn't yet set a light mode color so you get white on light.

We don't have a 'default text color' and none of the existing palette
entries fit, so I talked to Emma and decided to create a new variable.

Unfortunately, even though this is a 'correct' change, setting it right
now has the reverse problem: anything that is still missing light mode
support will see the default text color change from white to dark and
have problems on dark backgrounds. A quick test showed the following:

- some forms (#7214)
- #7727
- #7728

So we should not move this out of draft until most of these are fixed,
and definitely not immediately before a release. In the meantime, it's
handy to set this to something really obvious (e.g. amber) to see what
still needs light mode support.

Fixes #7724.

Signed-off-by: Tim deBoer <git@tdeboer.ca>