chore: configmap / secret icon
Adds the icon from design for Kubernetes configmaps and secrets.

This is the first icon we have without a 'solid' property (we decided
it looks better without). Svelte v5 has fixed the 'unknown property'
warning that would be triggered from StatusIcon, but I noticed
EmptyScreen is unnecessarily setting solid when that's the default for
all of our icons, so cleaning that up too.

Fixes #7680.

Signed-off-by: Tim deBoer <git@tdeboer.ca>