chore: use duration column for pod age
Same change as #7725, but for pods. Use the table duration column instead
of custom pod age. This cleans up the source and (by removing the 'if
running' there is no more age disappearing.

There's a separate issue that user might prefer 'time since started', but
that is unrelated.

Fixes #4972.
Fixes #2341.

Signed-off-by: Tim deBoer <git@tdeboer.ca>