chore: add GPU label for nodes with GPUs & adjust icons
### What does this PR do?

Adds the "GPU" label to indicate that a node has a GPU by checking to
see if Intel, Nvidia or AMD is present. This is useful as it provides
valuable information if a node is GPU capable or not.

This commit also adjusts the icons that represent a Node. I had
previously set it as microchip, but it should be "server" looking, as a
GPU is better indicated as a microchip / something that is added to a
server.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop/issues/7778

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

Add a "GPU" based node.. may be a bit difficult as you will have to
   follow a guide similar to
   https://docs.k3s.io/advanced#nvidia-container-runtime-support

However, tests currently cover this very niche edge-case.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>