chore: remove undefined for node in pods list
### What does this PR do?

If a node is not yet allocated, it will show as 'undefined' in the page.
Adds a check to make sure that we will only show the node name when
available.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

Closes https://github.com/containers/podman-desktop/issues/7780

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

Current tests cover issue, no point adding test for 'checking for
undefined'?

Signed-off-by: Charlie Drage <charlie@charliedrage.com>