chore: light mode support for statusbar
The status bar had hardcoded colors that weren't in the color palette.
This adds colors to the registry for them with the closest colors
(bg #302251 -> purple-900, hover-bg #4d3782 to purple-800) and uses
the variables.

This is the third time we're adding a notification dot color into the
registry and the previous two don't appear to be used yet, so I'm
separating it out into a single color, with the expectation that at
some point the notification dot should be a shared component like
Button or Label.

Signed-off-by: Tim deBoer <git@tdeboer.ca>