chore: light status icon default text color
The status icon is currently picking up the default text (fill) color from the
parent. This is fine at the moment while the default is white, but once we set
the default text color in light mode to black (see PR #7729) the fill color
would change to black.

This provides a new status color to use for the icons so that (as per design)
they use their own color and do not pick up the default/parent text color.

I wasn't 100% sure it should be a '--pd-status' color since this isn't a
status color per se like running or stopped, but it's much more obvious that
it is related to the other status colors.

Related to #7729. Fixes #7728.

Signed-off-by: Tim deBoer <git@tdeboer.ca>