chore: light error and warning colors
Replaces the hard-coded error and warning colors with values from the registry
and uses them for ErrorMessage, WarningMessage, TaskManagerItem, MessageBox, and
Donut.

I wasn't sure what to call this set as they are not quite status; went with state.
The MessageBox had state colors already defined for modals, but IMHO these should
be 'global' so I replaced them. These colors might need to be refined for light
and there are more uses of red-* and green-*, but this felt like enough to review
and get the color vars in.

Fixes #7936.

Signed-off-by: Tim deBoer <git@tdeboer.ca>