chore: consistent layout when >10 contexts
When we fixed #7657 it changed the layout of the Preferences > Kubernetes
page, which wasn't expecting contexts with no state. This just moves the
'{#if $kubernetesContextsState' down one step and adds an 'UNKNOWN' state
when a context has no status. Looks much bigger due to reformatting.

While taking the screenshot it was distracting that the right column beside
'server' didn't line up vertically with the others, so I found and removed
the inconsistent and unnecessary 'pl-0.5' too.

Fixes #7934.

Signed-off-by: Tim deBoer <git@tdeboer.ca>