chore: adjust search and tab bar width
Reduces the width of the search bar to a fixed w-72 (18rem) width, as before
it was switching between 22 and 34rem depending on the width of the window.

Adds a margin to the tab bar so that it is the same width as the table.

While I was doing this I noticed that the table is full width, but every row
has mx-5. Moved this to the top level so that it is consistent, and obvious
at the top level. IMHO table shouldn't have a built-in margin at all, but
that's an issue for another day (#7952).

Fixes #7949.
Fixes #7950.

Signed-off-by: Tim deBoer <git@tdeboer.ca>