chore: dialog component
We have a lot of inconsistencies in our dialogs - they all use Modal now, but
as shown in #7465 the header, content, and button bar are often styled
differently. As proposed this creates a Dialog component (based on what was
in MessageBox, which is the approved design) that has a fixed header (title
bar and close button), a padded content area, and a right-aligned bottom
button bar.

I've used the new component for MessageBox, SendFeedback, and RenameImageModal
(Edit Image) so that there is something to review; if this goes in we can then
reuse it for other dialogs.

I did _not_ attempt to change the styling within the content area or buttons of
these dialogs except where it was required to adopt the Dialog. I did make the
max content height a little bigger in order to fit the feedback form without
scrollbars.

First part of #7465.

Signed-off-by: Tim deBoer <git@tdeboer.ca>