chore: reusable base page
Provides a single base layout page that has enough options to support the current
FormPage and DetailsPages.

This is enough that both Help and Troubleshooting pages can use this new layout
instead of FormPage. This means they need to handle the close directly (maybe
it's worth another page in the middle to handle this for both?), but it means we
can remove the showBreadcrumb option from FormPage to start to clean it up.

First part of #7757. (NavPage, SettingsPage, or merging in any other pages is not
attempted here)

Fixes #7739.

Signed-off-by: Tim deBoer <git@tdeboer.ca>