chore: provide default text color
We have bg colors defined for every type of page, and several vars for
headers and other text colors in various forms. However, there is no 'default
text color' defined to catch any general text that isn't quite one of these
cases or isn't explicitly styled. As a result, the index.html of Podman
Desktop and extensions like AI Lab and Bootc still contain a 'text-white'.

We're not quite ready to change the default for Podman Desktop itself (see
PR #7729), but the missing variable stops extensions from doing it too.
Breaking this off into a separate PR to define the variable now, and we'll
rebase #7729 to use it in PD later.

Signed-off-by: Tim deBoer <git@tdeboer.ca>