chore: light mode for dashboard notifications
Adds standard light mode colors to the notification area. Makes the icon have the
same color as the notification dot.

Notifications weren't consistent with the Learning Center (which is the newer/more
recent design) so I did some minor things to make it more consistent: rounded
corners on the cards, header bigger, left, and left-justified.

Fixes #7998.

Signed-off-by: Tim deBoer <git@tdeboer.ca>