chore: default text color using variable
Replacement for #7729.

We should not hardcode the default text colour to white. Although it rarely shows
up in practice, it's better for it to match the current theme so that you don't
end up with white on white (e.g. see #7708)

There were a number of blockers before, because we didn't have enough light mode
done. However, we're only a few issues from the end and all major screens have
either been done or are imminent, so time to make this change too.

Fixes #7724.

Signed-off-by: Tim deBoer <git@tdeboer.ca>