chore: reduce vertical bump on modal dialogs
Our 'centered' dialogs aren't really centered - we bump them up vertically a little
bit because this helps draw attention and looks nice visually. However, when the
window is near it's smallest and the dialog is at its biggest, we can run into
problems where the top of the dialog goes 'underneath' the titlebar.

This reduces the vertical bump from 20% to 5%. This still gives the subtle effect,
but dialogs will never have this problem at our default window size. We could get
more clever and check the window size, reduce the maximum height of the dialog, or
try to fix fully by changing z-index and possibly adding window scrollbars if things
still don't fit, but this is way simpler and should solve >99% of the time.

Fixes #8091.

Signed-off-by: Tim deBoer <git@tdeboer.ca>