feat: add manifest api for pushing
### What does this PR do?

Took a while and found a bug regarding trying to push, but here is the
API for manifest pushing.

* Adds the API needed to push a manifest image
* Needs credentials, so does base64 encoding of credentials to API calls
* Requires ?all=true to work correctly.

There are other "optional" paramters (see:
https://docs.podman.io/en/latest/_static/api.html#tag/manifests/operation/ManifestPushLibpod), but we do not need to add them as
they are a not needed at the moment.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A, API.

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Part of https://github.com/containers/podman-desktop/issues/8057

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

Signed-off-by: Charlie Drage <charlie@charliedrage.com>