chore: light mode for troubleshooting
Replaces all hard-coded colors in the troubleshooting pages with color variables.
The two 'log' screens should probably be replaced with terminals so that we get
consistent styling, but terminal light mode/color vars are not there yet, so I've
given them basic card backgrounds for now.

When updating the copy to clipboard button I noticed that the EmptyScreen still
had hard-coded colors, so I fixed this and made the button in troubleshooting
match (link Button).

Signed-off-by: Tim deBoer <git@tdeboer.ca>