feat: basic light mode for monaco editors
The Monaco editor currently uses the vs-dark theme with a custom hard-coded
background color. This updates it to use the Monaco built-in light or dark
theme based on the current appearance, and uses our theme's terminal background
color for consistency.

We may want to support full customization of Monaco in the future, but there
are lots of possible colors (variables, constants, comments, etc.) so it is
not trivial. This just does the basic step to keep what we have today with a
consistent look in light mode.

Signed-off-by: Tim deBoer <git@tdeboer.ca>