chore: light mode for docker desktop extensions (and more)
In order to set the correct prefers-color-scheme for Docker Desktop extensions
in <webview>s we need to set the Electron native theme, from main. This is
probably a good thing to do anyway since it sets the preferred global color
scheme correctly, and has the side-effect of also setting the light/dark mode
for the developer tools.

Fixes #8217.

Signed-off-by: Tim deBoer <git@tdeboer.ca>