chore: add try/catch to ssh connection
### What does this PR do?

Right now we have no error out put when we cannot initially connect, for
example, having a bad key being passed in.

We also do not check for any errors on the initial connection either at
the moment.

This adds a try / catch so we can make sure that we at least put the
error output in the console and diagnose what is going on.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

N/A, adds try/catch.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>