chore: allow simple column to wrap to 2 lines
I had a change to make service port a special column that could wrap to two lines,
but then I figured why not all simple columns? If text is too wide to fit into a
column this will cause it to wrap to two lines. Longer text will still use ellipsis.
(we could go to 3 or 4 lines, but allowing twice as much text seemed like a
reasonable balance that doesn't affect the rest of the table layout)

Having multiple service ports are the most obvious way to trigger this, but on
very narrow columns something like "2 months" will be wrapped to:
"2
months"
instead of:
"2 mo..."

Fixes #5442.

Signed-off-by: Tim deBoer <git@tdeboer.ca>