fix: update native theme directly on config change
PR #8289 set the native theme so that extensions render correctly, but once we
set the native theme, window.matchMedia('(prefers-color-scheme: dark)') just
returns what we set the native theme to - i.e. if system is in light mode and
you explicitly set our config value to dark mode and then to system, we will
ask the system what it prefers and it says 'dark', because that's what we told
it to do last.

The fix is to set the native theme directly (and immediately) based on the value
of our configuration setting. That way when we call isDarkMode() later on if
we're in system mode we will get the correct value from prefers-color-scheme.

Since we don't need nativeTheme exposed to renderer anymore I'm removing it.

FWIW I also tried updating native theme on config change in Appearance's
onDidChangeConfigurationCallback. It works fine, but overall this felt simpler
and more direct.

Fixes #8477.

Signed-off-by: Tim deBoer <git@tdeboer.ca>