chore: reduce header width
Even after removing Core Values (#8964), the Podman Desktop title is often cut
off when I open our website in a browser. You have to have a large or full-
screen browser to see the whole Podman Desktop title.

Comparing with other sites, our logo + title is a bit wide (not much we can
do about the name) and our header font sizes are also on the high side. This
just reduces the font size a little and slightly increases the reactive width
where it jumps between regular and large fonts so that Podman Desktop is
usually visible.

Fixes #5761.

Signed-off-by: Tim deBoer <git@tdeboer.ca>