fix: show correct empty screens when filtering catalog
See #9056, when searching the extension catalog many of the empty screens
were incorrect. This moves the FilteredEmptyScreen down into the installed
and catalog sections so that it is displayed correctly for each screen, and
reuses the existing showEmptyScreen property on CatalogExtensionList to
avoid showing duplicate empty screens.

Fixes three issues:

- When search term isn't in installed or catalog, show (only) the
  FilteredEmptyScreen in both. (was duplicate empty screen in catalog)
- When search term is in installed but not in catalog, show the
  FilteredEmptyScreen in catalog. (was internet connection empty screen)
- When search term is in catalog but not in installed, show the
  FilteredEmptyScreen in installed. (was blank screen)

Fixes #9056.

Signed-off-by: Tim deBoer <git@tdeboer.ca>