feat: dropdown component
As per the last time this item was discussed, I spent a while trying to use
svelte-select and a few other off-the-shelf select components to work, but
ultimately they all had the same problem as we had with pulling images: even
after tying to style them, I couldn't get any of them to fit into our current
styling well enough that I was comfortable submitting a PR.

Two other minor things to note:
- Most components used some form of item[] instead of child option elements.
  This might be useful in the future (and could be added here) but would have
  required more migration up front.
- We don't have any immediate plans to expose this component in the UI library,
  but using a third-party library would have excluded that possibility.

So, here it is instead: I used the code from the Typeahead component, mixing
in bits from Tooltip and other places to create a custom styled Dropdown
component. I tried to keep required migration to a minimum, but some tests
needed to be updated and I switched onInput to svelte 5 style.

Added to UI package and added a basic storybook story to test. There are >15
places in renderer where we use an HTML select which would be migrated in
future PRs.

Signed-off-by: Tim deBoer <git@tdeboer.ca>