chore: redirect kubernetes page to /kubernetes/nodes
### What does this PR do?

When clicking on the navigation, it will take you to `kubernetes/nodes`
instead of `/kubernetes` which may lead to a blank page.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes containers#9189

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

Click on Kubernetes and it'll go to /nodes now.

- [X] Tests are covering the bug fix or the new feature

Signed-off-by: Charlie Drage <charlie@charliedrage.com>