fix: exited quick pick should resolve to undefined
The documentation for quick picks says: `@return A promise that resolves to the
selected items or 'undefined'` but instead it rejected the promise as 'no item'
if the user escapes or exits the popup.

This adds an explicit undefined back from the renderer and resolves to
undefined to match the spec.

Fixes #9167.

Signed-off-by: Tim deBoer <git@tdeboer.ca>