chore: use dropdown component in Kube terminal
Switches the select in the Kube terminal to a Dropdown.

The select control automatically selects the first value if none is provided.
Dropdown wasn't doing that so I added the appropriate line, hoping it is
small enough to justify including it here.

Remove redundant aria-labelledby (label already points to control).

Part of #6199.

Signed-off-by: Tim deBoer <git@tdeboer.ca>