fix: use option property to ensure options is treated as state
See #9917 for background. Svelte reuses component instances whenever possible. The
problem here is that the Dropdown options are being set via child option elements,
created via buildOptions. When you switch pages all of the explicitly set properties
and state are applied, but since options is not set as a property, nor state, it is
not applied.

The simplest fix in this case is using the options parameter directly, which we do
in other cases and causes Svelte to correctly treat it as state.

This is only avoiding the problem though, so I've opened #9952 for future follow up.
I've checked and we have no other child options where this would currently happen,
but the real fix is possibly a Svelte fix, removing child option support, or finding
a fix directly in Dropdown.

Fixes #9917.

Signed-off-by: Tim deBoer <git@tdeboer.ca>