fix: remove failing part of test
The test was bad - checking the text content of a select or dropdown _always_
passes when the value you're looking for is also a child option. Switching to
not use children exposes the invalid test.

This just removes the invalid line, which reverts this test to exactly where it
was before we added Dropdown, just checking that the correct component is used.
The correct initial selection behaviour is checked in the following test.

Signed-off-by: Tim deBoer <git@tdeboer.ca>