refactor: adding a check for the KUBECONFIG variable
This validation of the existence of the KUBECONFIG
variable set in the operating system aims to inform
the use through the console that this could lead
to an unexpected behavior in some scenarios, like
trying to install a Kubernetes cluster and failing
because Podman Desktop cannot use or update the
default ~/.kube/config file. The issue #9983 describes
this better with an example of such situation. Although
this not correct the variable value, at least informs
the user of its value, allowing the user to take the
proper action to fix it.

Closes #9983

Signed-off-by: Alexon Oliveira <alexon@redhat.com>
Signed-off-by: Alexon Oliveira <alolivei@redhat.com>